chore(release): version 20250103.1
**Release Date:** January 3, 2026

**Highlights:**
- Claude model issues resolved (infinite loops fixed)
- Simplified agent architecture (removed LoopDetector.swift)
- Fine-grained CFG scale control (decimal values like 5.9, 7.5)
- GitHub Copilot API bugs corrected
- Smarter tool result previews (8KB immediate visibility)
- HTML release notes format

**Bug Fixes:**
- Claude no longer loops infinitely on complex workflows
- GitHub Copilot API errors eliminated
- Tool results now show actual content
- Auto-update version comparison corrected

**Technical Changes:**
- Decimal CFG scale: ConversationModel.sdGuidanceScale Int → Float
- ImageGenerationTool accepts Int/Float/Double for backward compatibility
- UI updated with 0.1 step increments and decimal display
- ChatWidget.performMainChatViewAppear() extracted
- Removed LoopDetector.swift (575 lines)
- AgentOrchestrator simplified (-127 lines, -22%)

**Files Modified:**
- Info.plist: v20251230.1 → v20250103.1
- Resources/whats-new.json: User-friendly release notes
- Sources/ConversationEngine/ConversationModel.swift
- Sources/StableDiffusionIntegration/ImageGenerationTool.swift
- Sources/UserInterface/Chat/ChatWidget.swift

+++ Sources/ConversationEngine/ConversationModel.swift
@@ -64,7 +64,7 @@ public struct ConversationSettings: Codable, Sendable {
     /// Stable Diffusion parameters
     public var sdNegativePrompt: String
     public var sdSteps: Int
-    public var sdGuidanceScale: Int
+    public var sdGuidanceScale: Float
     public var sdScheduler: String
     public var sdSeed: Int
     public var sdUseKarras: Bool  /// Karras sigma schedule for DPM++
@@ -106,7 +106,7 @@ public struct ConversationSettings: Codable, Sendable {
         draftMessage: String = "",
         sdNegativePrompt: String = "",
         sdSteps: Int = 25,
-        sdGuidanceScale: Int = 8,
+        sdGuidanceScale: Float = 8.0,
         sdScheduler: String = "dpm++",
         sdSeed: Int = -1,
         sdUseKarras: Bool = true,
@@ -228,7 +228,7 @@ public struct ConversationSettings: Codable, Sendable {
         /// Stable Diffusion parameters with defaults for backward compatibility.
         sdNegativePrompt = try container.decodeIfPresent(String.self, forKey: .sdNegativePrompt) ?? ""
         sdSteps = try container.decodeIfPresent(Int.self, forKey: .sdSteps) ?? 25
-        sdGuidanceScale = try container.decodeIfPresent(Int.self, forKey: .sdGuidanceScale) ?? 8
+        sdGuidanceScale = try container.decodeIfPresent(Float.self, forKey: .sdGuidanceScale) ?? 8.0
         sdScheduler = try container.decodeIfPresent(String.self, forKey: .sdScheduler) ?? "dpm++"
         sdSeed = try container.decodeIfPresent(Int.self, forKey: .sdSeed) ?? -1
         sdUseKarras = try container.decodeIfPresent(Bool.self, forKey: .sdUseKarras) ?? true
